WIP Refactor of bindings
