+++ packages/wrangler/src/config/environment.ts
@@ -26,6 +26,8 @@ export type Route =
 	| ZoneNameRoute
 	| CustomDomainRoute;
 
+import { BindingTypesNonInheritable } from "./../bindings/bindings"
+
 /**
  * The `EnvironmentInheritable` interface declares all the configuration fields for an environment
  * that can be inherited (and overridden) from the top-level environment.
@@ -267,17 +269,6 @@ interface EnvironmentInheritable {
 	logpush: boolean | undefined;
 }
 
-export type DurableObjectBindings = {
-	/** The name of the binding used to refer to the Durable Object */
-	name: string;
-	/** The exported class name of the Durable Object */
-	class_name: string;
-	/** The script where the Durable Object is defined (if it's external to this worker) */
-	script_name?: string;
-	/** The service environment of the script_name to bind to */
-	environment?: string;
-}[];
-
 /**
  * The `EnvironmentNonInheritable` interface declares all the configuration fields for an environment
  * that cannot be inherited from the top-level environment, and must be defined specifically.
@@ -285,7 +276,7 @@ export type DurableObjectBindings = {
  * If any of these fields are defined at the top-level then they should also be specifically defined
  * for each named environment.
  */
-interface EnvironmentNonInheritable {
+interface EnvironmentNonInheritable extends BindingTypesNonInheritable {
 	/**
 	 * A map of values to substitute when deploying your worker.
 	 *
@@ -306,22 +297,6 @@ interface EnvironmentNonInheritable {
 	 * @nonInheritable
 	 */
 	vars: { [key: string]: unknown };
-
-	/**
-	 * A list of durable objects that your worker should be bound to.
-	 *
-	 * For more information about Durable Objects, see the documentation at
-	 * https://developers.cloudflare.com/workers/learning/using-durable-objects
-	 *
-	 * NOTE: This field is not automatically inherited from the top level environment,
-	 * and so must be specified in every named environment.
-	 *
-	 * @default `{bindings:[]}`
-	 * @nonInheritable
-	 */
-	durable_objects: {
-		bindings: DurableObjectBindings;
-	};
 
 	/**
 	 * These specify any Workers KV Namespaces you want to
